fixed tests
